// typescript/packages/objects/generated/proto/objects/v1/objects_pb.d.ts
export declare type BucketRef = Message<"acyclic.objects.v1.BucketRef"> & {
  /**
   * @generated from field: string bucket_id = 1;
   */
  bucketId: string;

  /**
   * @generated from field: string name = 2;
   */
  name: string;
};
// ...
export declare type ObjectMetadata = Message<"acyclic.objects.v1.ObjectMetadata"> & {
  /**
   * @generated from field: string content_type = 1;
   */
  contentType: string;

  /**
   * @generated from field: map<string, string> user = 2;
   */
  user: { [key: string]: string };

  /**
   * @generated from field: string content_encoding = 3;
   */
  contentEncoding: string;

  /**
   * @generated from field: string cache_control = 4;
   */
  cacheControl: string;

  /**
   * @generated from field: string content_disposition = 5;
   */
  contentDisposition: string;

  /**
   * @generated from field: string content_language = 6;
   */
  contentLanguage: string;

  /**
   * @generated from field: optional int64 expires_unix_seconds = 7;
   */
  expiresUnixSeconds?: bigint | undefined;
};

/**
 * Describes the message acyclic.objects.v1.ObjectMetadata.
 * Use `create(ObjectMetadataSchema)` to create a new message.
 */
export declare const ObjectMetadataSchema: GenMessage<ObjectMetadata>;

/**
 * @generated from message acyclic.objects.v1.Preconditions
 */
export declare type Preconditions = Message<"acyclic.objects.v1.Preconditions"> & {
  /**
   * @generated from oneof acyclic.objects.v1.Preconditions.condition
   */
  condition: {
    /**
     * @generated from field: bool if_absent = 1;
     */
    value: boolean;
    case: "ifAbsent";
  } | {
    /**
     * @generated from field: string if_match = 2;
     */
    value: string;
    case: "ifMatch";
  } | {
    /**
     * @generated from field: string if_version = 3;
     */
    value: string;
    case: "ifVersion";
  } | { case: undefined; value?: undefined };
};

/**
 * Describes the message acyclic.objects.v1.Preconditions.
 * Use `create(PreconditionsSchema)` to create a new message.
 */
export declare const PreconditionsSchema: GenMessage<Preconditions>;

/**
 * @generated from message acyclic.objects.v1.MutationIdentity
 */
export declare type MutationIdentity = Message<"acyclic.objects.v1.MutationIdentity"> & {
  /**
   * @generated from field: string idempotency_key = 1;
   */
  idempotencyKey: string;
};
// ...
export declare type PutObjectHeader = Message<"acyclic.objects.v1.PutObjectHeader"> & {
  /**
   * @generated from field: acyclic.objects.v1.BucketRef bucket = 1;
   */
  bucket?: BucketRef | undefined;

  /**
   * @generated from field: string object_key = 2;
   */
  objectKey: string;

  /**
   * @generated from field: acyclic.objects.v1.ObjectMetadata metadata = 3;
   */
  metadata?: ObjectMetadata | undefined;

  /**
   * @generated from field: acyclic.objects.v1.Preconditions preconditions = 4;
   */
  preconditions?: Preconditions | undefined;

  /**
   * @generated from field: acyclic.objects.v1.MutationIdentity mutation = 5;
   */
  mutation?: MutationIdentity | undefined;
};

/**
 * Describes the message acyclic.objects.v1.PutObjectHeader.
 * Use `create(PutObjectHeaderSchema)` to create a new message.
 */
export declare const PutObjectHeaderSchema: GenMessage<PutObjectHeader>;

/**
 * @generated from message acyclic.objects.v1.PutObjectRequest
 */
export declare type PutObjectRequest = Message<"acyclic.objects.v1.PutObjectRequest"> & {
  /**
   * @generated from oneof acyclic.objects.v1.PutObjectRequest.frame
   */
  frame: {
    /**
     * @generated from field: acyclic.objects.v1.PutObjectHeader header = 1;
     */
    value: PutObjectHeader;
    case: "header";
  } | {
    /**
     * @generated from field: bytes body = 2;
     */
    value: Uint8Array;
    case: "body";
  } | { case: undefined; value?: undefined };
};

/**
 * Describes the message acyclic.objects.v1.PutObjectRequest.
 * Use `create(PutObjectRequestSchema)` to create a new message.
 */
export declare const PutObjectRequestSchema: GenMessage<PutObjectRequest>;

/**
 * @generated from message acyclic.objects.v1.ObjectVersion
 */
export declare type ObjectVersion = Message<"acyclic.objects.v1.ObjectVersion"> & {
  /**
   * @generated from field: string version_id = 1;
   */
  versionId: string;

  /**
   * @generated from field: string etag = 2;
   */
  etag: string;

  /**
   * @generated from field: uint64 size = 3;
   */
  size: bigint;

  /**
   * @generated from field: bool delete_marker = 4;
   */
  deleteMarker: boolean;

  /**
   * @generated from field: acyclic.objects.v1.ObjectMetadata metadata = 5;
   */
  metadata?: ObjectMetadata | undefined;

  /**
   * @generated from field: google.protobuf.Timestamp created_at = 6;
   */
  createdAt?: Timestamp | undefined;
};
// ...
export declare type GetObjectResponse = Message<"acyclic.objects.v1.GetObjectResponse"> & {
  /**
   * @generated from oneof acyclic.objects.v1.GetObjectResponse.frame
   */
  frame: {
    /**
     * @generated from field: acyclic.objects.v1.ObjectVersion version = 1;
     */
    value: ObjectVersion;
    case: "version";
  } | {
    /**
     * @generated from field: bytes body = 2;
     */
    value: Uint8Array;
    case: "body";
  } | { case: undefined; value?: undefined };
};

/**
 * Describes the message acyclic.objects.v1.GetObjectResponse.
 * Use `create(GetObjectResponseSchema)` to create a new message.
 */
export declare const GetObjectResponseSchema: GenMessage<GetObjectResponse>;
// ...
export declare type UploadPartHeader = Message<"acyclic.objects.v1.UploadPartHeader"> & {
  /**
   * @generated from field: acyclic.objects.v1.BucketRef bucket = 1;
   */
  bucket?: BucketRef | undefined;

  /**
   * @generated from field: string object_key = 2;
   */
  objectKey: string;

  /**
   * @generated from field: string upload_id = 3;
   */
  uploadId: string;

  /**
   * @generated from field: uint32 part_number = 4;
   */
  partNumber: number;

  /**
   * @generated from field: acyclic.objects.v1.MutationIdentity mutation = 5;
   */
  mutation?: MutationIdentity | undefined;
};

/**
 * Describes the message acyclic.objects.v1.UploadPartHeader.
 * Use `create(UploadPartHeaderSchema)` to create a new message.
 */
export declare const UploadPartHeaderSchema: GenMessage<UploadPartHeader>;

/**
 * @generated from message acyclic.objects.v1.UploadPartRequest
 */
export declare type UploadPartRequest = Message<"acyclic.objects.v1.UploadPartRequest"> & {
  /**
   * @generated from oneof acyclic.objects.v1.UploadPartRequest.frame
   */
  frame: {
    /**
     * @generated from field: acyclic.objects.v1.UploadPartHeader header = 1;
     */
    value: UploadPartHeader;
    case: "header";
  } | {
    /**
     * @generated from field: bytes body = 2;
     */
    value: Uint8Array;
    case: "body";
  } | { case: undefined; value?: undefined };
};

/**
 * Describes the message acyclic.objects.v1.UploadPartRequest.
 * Use `create(UploadPartRequestSchema)` to create a new message.
 */
export declare const UploadPartRequestSchema: GenMessage<UploadPartRequest>;
// ...
export declare type ErrorDetail = Message<"acyclic.objects.v1.ErrorDetail"> & {
  /**
   * @generated from field: acyclic.objects.v1.ErrorCode code = 1;
   */
  code: ErrorCode;

  /**
   * @generated from field: string request_id = 2;
   */
  requestId: string;
};

/**
 * Describes the message acyclic.objects.v1.ErrorDetail.
 * Use `create(ErrorDetailSchema)` to create a new message.
 */
export declare const ErrorDetailSchema: GenMessage<ErrorDetail>;
// ...
export enum ErrorCode {
  /**
   * @generated from enum value: ERROR_CODE_UNSPECIFIED = 0;
   */
  UNSPECIFIED = 0,

  /**
   * @generated from enum value: ERROR_CODE_INVALID_ARGUMENT = 1;
   */
  INVALID_ARGUMENT = 1,

  /**
   * @generated from enum value: ERROR_CODE_NOT_FOUND = 2;
   */
  NOT_FOUND = 2,

  /**
   * @generated from enum value: ERROR_CODE_ALREADY_EXISTS = 3;
   */
  ALREADY_EXISTS = 3,

  /**
   * @generated from enum value: ERROR_CODE_PRECONDITION_FAILED = 4;
   */
  PRECONDITION_FAILED = 4,

  /**
   * @generated from enum value: ERROR_CODE_IDEMPOTENCY_MISMATCH = 5;
   */
  IDEMPOTENCY_MISMATCH = 5,

  /**
   * @generated from enum value: ERROR_CODE_TOKEN_EXPIRED = 6;
   */
  TOKEN_EXPIRED = 6,

  /**
   * @generated from enum value: ERROR_CODE_QUOTA_EXCEEDED = 7;
   */
  QUOTA_EXCEEDED = 7,

  /**
   * @generated from enum value: ERROR_CODE_UNSUPPORTED = 8;
   */
  UNSUPPORTED = 8,

  /**
   * @generated from enum value: ERROR_CODE_UNAVAILABLE = 9;
   */
  UNAVAILABLE = 9,
}